+++ docs/scripts/app.js
@@ -1,0 +1,18 @@
+/**
+ * Generated header by C0ZEN for 21-days project
+ * Generated file app on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 29/12/2017
+ * Time: 23:31
+ * Version: 1.0.0
+ */
+(function (angular) {
+	'use strict';
+
+	angular
+		.module('21days', [
+			'720kb.datepicker'
+		]);
+
+})(window.angular);