+++ docs/scripts/main.controller.js
@@ -1,0 +1,25 @@
+/**
+ * Generated header by Geoffrey Testelin for 21-days project
+ * Created with: generator-cozen-angular
+ *
+ * Created by: Geoffrey Testelin
+ * Date: 29/12/2017
+ * Time: 23:34
+ * Version: 0.1.4
+ */
+(function (angular) {
+	'use strict';
+
+	angular
+		.module('21days')
+		.controller('mainController', mainController);
+
+	mainController.$inject = [];
+
+	function mainController() {
+		let vm = this;
+
+		vm.methods = {};
+	}
+
+})(window.angular);