+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Angular app
+- Angular main controller
+- Generator-cozen-angular config
+
+### Changed
+- Main styles
 
 ## [0.1.4]
 ### Added